requirements.txt
attrs==23.2.1.dev0
autocommand==2.2.2
btrfsutil==6.12
build==1.2.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
editables==0.5
fastjsonschema==2.21.1
filelock==3.16.1
hatchling==1.25.0
hyprshade==3.2.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==7.2.1
installer==0.7.0
jaraco.collections==5.0.1
jaraco.context==5.3.0
jaraco.functools==4.1.0
jaraco.text==4.0.0
lark==1.2.2
lxml==5.3.0
Mako==1.3.8.dev0
Markdown==3.7
MarkupSafe==2.1.5
meson==1.6.1
more-itertools==10.5.0
numpy==2.2.1
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
pillow_heif==0.21.0
platformdirs==4.3.6
pluggy==1.5.0
poetry-core==1.9.1
pycairo==1.27.0
pycparser==2.22
Pygments==2.18.0
PyGObject==3.50.0
pyproject_hooks==1.2.0
pywal==3.3.0
requests==2.32.3
screeninfo==0.8.1
setuptools==75.2.0
tqdm==4.67.1
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
urllib3==1.26.20
waypaper==2.4
wheel==0.45.0
yt-dlp==2024.12.23
zipp==3.21.0
